backend/requirements.txt
# Backend-specific dependencies
# (jobscout core is installed via pyproject.toml in Dockerfile)

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
apscheduler>=3.10.0
python-dotenv>=1.0.0
python-docx>=1.1.0